+++ R/data.R
@@ -1,0 +1,8 @@
+#' varianceStabilizingTransformation Object from Test Data
+#'
+#' An example of the output of varianceStabilizingTransformation to
+#' show different visualization code
+#'
+#' @format ## `bcbio_vsd_data`
+#' VSD object from DESeq2 data
+"bcbio_vsd_data"