+++ inst/rmarkdown/templates/cosmx/skeleton/information.R
@@ -1,0 +1,6 @@
+# info params
+project = "name_hbcXXXXX"
+PI = 'person name'
+experiment = 'short description'
+aim = 'short description'
+analyst = 'person in the core'