+++ inst/rmarkdown/templates/rnaseq/skeleton/QC/params_qc_nf-core-example.R
@@ -1,0 +1,9 @@
+# info params
+
+# Example data: COMMENT THESE LINE IF YOU ARE USING YOUR DATA
+metadata_fn='https://raw.githubusercontent.com/bcbio/bcbioR-test-data/devel/rnaseq/nf-core/coldata.csv'
+se_object=url('https://raw.githubusercontent.com/bcbio/bcbioR-test-data/devel/rnaseq/nf-core/star_salmon/salmon.merged.gene_counts.rds')
+# This folder is in the output directory inside multiqc folder
+multiqc_data_dir='https://raw.githubusercontent.com/bcbio/bcbioR-test-data/devel/rnaseq/nf-core/multiqc/star_salmon/multiqc-report-data/'
+# This file is inside the genome folder in the output directory
+gtf_fn='https://raw.githubusercontent.com/bcbio/bcbioR-test-data/main/devel/nf-core/genome/genome.filtered.gtf.gz'