+++ inst/rmarkdown/templates/rnaseq/skeleton/QC/params_qc_nf-core.R
@@ -1,5 +1,11 @@
 # info params
 
-metadata_fn='https://raw.githubusercontent.com/bcbio/bcbioR-test-data/main/rnaseq/nf-core/coldata.csv'
-se_object=url('https://raw.githubusercontent.com/bcbio/bcbioR-test-data/main/rnaseq/nf-core/salmon.merged.gene_counts.rds')
-multiqc_data_dir='https://raw.githubusercontent.com/bcbio/bcbioR-test-data/main/rnaseq/nf-core/multiqc-report-data/'
+# Your data
+# This is the file used to run nf-core or compatible to that
+metadata_fn='/Path/to/metadata/meta.csv'
+# This file is inside star_salmon/ folder
+se_object='/path/to/nf-core/output/star_salmon/salmon.merged.gene_counts.rds'
+# This folder called "multiqc_report_data" is inside the output directory star_salmon inside multiqc folder
+multiqc_data_dir='/path/to/nf-core/output/star_salmon/multiqc_report_data'
+# This file is inside the genome folder in the output directory, use this only for non-model organism
+# gtf_fn='/path/to/nf-core/output/genome/hg38.filtered.gtf'