+++ inst/rmarkdown/templates/rnaseq/skeleton/QC/run_markdown.R
@@ -1,10 +1,13 @@
 library(rmarkdown)
 
-rmarkdown::render("./inst/rmarkdown/templates/rnaseq/skeleton/QC/QC_nf-core.Rmd",
-                  output_dir = "./inst/rmarkdown/templates/rnaseq/skeleton/QC/", 
+# set directory to this file folder
+setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
+# example running with test data
+rmarkdown::render("QC_nf-core.Rmd",
+                  output_dir = ".",
                   clean = TRUE,
                   output_format = "html_document",
                   params = list(
-                    params_file = 'params_qc_nf-core.R',
+                    params_file = 'params_qc_nf-core-testdata.R',
                     project_file = '../information.R')
                   )