+++ NEWS.md
@@ -1,3 +1,16 @@
+# bcbioR 0.1.2
+
+* Add combatseq
+* Add draft templates for TEAseq, COSMX
+* Adapt templates to nf-core rnaseq
+* Fix when sample start by number
+* Fix when rRNA rate is missing
+* Add by sample plots in QC
+* Add function to check nfcore samplesheet
+* Add PCA with variance analysis
+* Minor details fixed in QC RNAseq report
+* Correct metric for rRNA and tRNA
+
 # bcbioR 0.1.1
 
 * Add page to github